fix leaks
