+++ Makefile
@@ -19,7 +19,7 @@ SRCS		=	$(patsubst %.c, $(SRC)/%.c,\
 				$(patsubst %.c, $(SRC)/Utils/%.c,\
 				lst_new.c lst_get.c lst_push.c \
 				lst_last.c lst_size.c lst_clear.c \
-				get_env.c) 
+				get_env.c lst_set.c) 
 
 
 OBJS		=	$(patsubst %.c, $(BUILD)/%.o, main.c) \
