+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: dohanyan <dohanyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/26 18:00:51 by dohanyan          #+#    #+#             */
-/*   Updated: 2023/06/02 21:33:28 by dohanyan         ###   ########.fr       */
+/*   Updated: 2023/06/03 16:59:51 by dohanyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void	history(void);
 
 t_list	*env_init(char **env);
 void	pwd(void);
-void cd(const char *path,t_list *env);
+void	cd(const char *path,t_list *env);
 void	env(t_list *env);
 
 /* Utils */
