+++ src/BuiltIns/env.c
@@ -6,7 +6,7 @@
 /*   By: dohanyan <dohanyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 15:10:00 by dohanyan          #+#    #+#             */
-/*   Updated: 2023/06/02 21:32:35 by dohanyan         ###   ########.fr       */
+/*   Updated: 2023/06/03 19:28:05 by dohanyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,4 @@ void env(t_list *env)
 		printf("%s\n",temp->joined);
 		temp = temp->next;
 	}
-	
 }
