+++ src/BuiltIns/pwd.c
@@ -6,7 +6,7 @@
 /*   By: dohanyan <dohanyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 17:01:08 by dohanyan          #+#    #+#             */
-/*   Updated: 2023/05/31 20:18:03 by dohanyan         ###   ########.fr       */
+/*   Updated: 2023/06/03 21:03:24 by dohanyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
